converted function forecastCard into a widgеt ForecastCard

+++ lib/models/weather_forecast_daily.dart
@@ -152,7 +152,7 @@ class WeatherList {
     }
     speed = json['speed'];
     deg = json['deg'];
-    gust = json['gust'];
+    gust = json['gust'].toDouble();
     clouds = json['clouds'];
     pop = json['pop'].toDouble();
   }
